+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Rename widget to component
 * Rename `.sfw` to `.sfc`
 * Rename controllers to initializers
+* Destroy argument for remove generated scaffold items [#11](https://github.com/sham-ui/sham-ui-cli/issues/11)
 
 ## 2.0.4
 * Fix `sham-ui-test-helpers` version in app template